add a static library + fix the CLEAN
Add a static library build directive, so as to ease link from third
party softwares: instead of having to indicate each object, the archive
alone is sufficient.

Fix the CLEAN directive of the high level Makefile to effectively remove
objects and the static library. The fix is minimal and do not modify the
existence of two separated directive in each subdirectories.