[feature] add first PoC of an IOH service
This adds a problem server that takes JSON queries through named pipes.
The server host an Analyzer logger and is able to serve objective function computation on existing problems.
The use of named pipes allows for very simple code, without managing states or network sockets (which can/will be managed by another service).
The implementation comes with a client example implemented in Python.

The JSON schemas of the protocol messages are provided.
Messages can be checked against the schemas via a side tool taking the form of a SingularityCE container (in order to avoid adding optional dependencies to IOH).

The commit adds two single-header sub-modules:
- cxxopts (for CLI argument management)
- Nlohmann's json parsing.